user增加major rank

+++ src/main/java/cn/edu/nju/mapper/MajorRankMapper.java
@@ -19,7 +19,7 @@ public interface MajorRankMapper {
 
     List<MajorRankVO> selectAll();
 
-    Integer selectRank(@Param("institude") String institude,@Param("major") String major);
+    Double selectRank(@Param("institude") String institude,@Param("major") String major);
 
     List<MajorRankVO> selectByInstitude(String institude);
